Fix g++ warnings

+++ libdirac_encoder/seq_compress.cpp
@@ -103,7 +103,7 @@ SequenceCompressor::SequenceCompressor( StreamPicInput* pin ,
     int yl_chroma = sparams.Yl() / y_chroma_fac;
      xpad_chroma = ypad_chroma = 0;  
     // The pic dimensions must be a multiple of 2^(transform depth). 
-    int tx_mul = static_cast<int>(std::pow(2.0, m_encparams.TransformDepth()));
+    int tx_mul = static_cast<int>(std::pow(2.0, (int)m_encparams.TransformDepth()));
 
     if ( xl_chroma%tx_mul != 0 )
         xpad_chroma = ( (xl_chroma/tx_mul)+1 ) *tx_mul - xl_chroma;
